fix: split type util function
