+++ packages/dataDisplay/useSortableTable/src/useSortableTable.tsx
@@ -1,47 +1,7 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
+import { isValidKeysTypeArray, isValidKeyType, isSubTypeArray } from "./utils";
 
 type SortableTableDataType<K extends keyof V, V extends Object> = Map<K, Array<V[K]>>;
-
-const isValidKeysType = <K extends keyof V, V extends Object>(data: V, keys: (string | number | Symbol)[]): keys is K[] => {
-    const dataKeys = Object.keys(data);
-    if (dataKeys.every((dataKey) => keys.includes(dataKey))) {
-        return true;
-    }
-    return false;
-}
-
-const isValidKeysTypeArray = <K extends keyof V, V extends Object>(data: V[], keys: (string | number | Symbol | K)[]): keys is K[] => data.every((d) => isValidKeysType<K, V>(d, keys));
-
-const isValidKeyType = <K extends keyof V, V extends Object>(data: V, key: string | number | Symbol): key is K => {
-    const dataKeys = Object.keys(data);
-    if (dataKeys.includes(key.toString())) {
-        return true;
-    }
-    return false;
-}
-
-const isSubType = <T, U extends T>(data: T): data is U => {
-    const subTypeData = data as U;
-
-    if (typeof data === "object" && data !== null && typeof subTypeData === "object" && subTypeData !== null) {
-
-        const originDataKey = Object.keys(data);
-        const subTypeDataKey = Object.keys(subTypeData);
-
-        if (originDataKey.length === subTypeDataKey.length) {
-            return originDataKey.every((key) => subTypeDataKey.includes(key));
-        }
-        return false;
-    }
-
-    if (typeof data === typeof subTypeData) {
-        return true;
-    }
-
-    return false;
-}
-
-const isSubTypeArray = <T, U extends T>(data: T[]): data is U[] => data.every((d) => isSubType<T, U>(d));
 
 const useSortableTable = <K extends keyof V, V extends Object>(data: V[]) => {
     const [convertedData, setConvertedData] = useState<SortableTableDataType<K, V> | undefined>(undefined);
@@ -134,7 +94,7 @@ const useSortableTable = <K extends keyof V, V extends Object>(data: V[]) => {
                 }                
                 obj[key] = targetMapArray[index];
                 return obj;
-            }, Object.assign({}))
+            }, {} as V)
         );
         setSortableData(sortedData);
     }, [convertedData, keys]);
